requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.36
transformers==4.46.3
datasets[audio]==3.1.0
accelerate==1.1.1
librosa==0.10.2.post1
wget==3.2
email-validator==2.3.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic-settings==2.0.3
pika==1.3.2

